+++ src/users/users.service.ts
@@ -1,7 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, UseGuards } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { AuthGuard } from 'src/auth/auth.guard';
+import { Request } from 'express';
 
 @Injectable()
 export class UsersService {
@@ -13,7 +15,7 @@ export class UsersService {
       let all = await this.prisma.user.findMany()
       return all
     } catch (error) {
-      return {message: error.message}
+      throw new BadRequestException({message: error.message})
     }
   }
 
@@ -22,23 +24,31 @@ export class UsersService {
       let one = await this.prisma.user.findFirst({where: {id}})
       return one
     } catch (error) {
-      return {message: error.message}
+      throw new BadRequestException({message: error.message})
     }
   }
 
-  async update(id: number, data: UpdateUserDto) {
+  async update(id: number, data: UpdateUserDto, req: Request) {
     try {
+      if(req['user'].role != "ADMIN" && req['user'].id == id){
+        throw new BadRequestException("You cannot updated othes account! Only ADMIN or SUPER-ADMIN can!")
+      }
       let update = await this.prisma.user.update({where: {id}, data})
+      return update
     } catch (error) {
-      return {message: error.message}
+      throw new BadRequestException({message: error.message})
     }
   }
 
-  async remove(id: number) {
+  async remove(id: number, req: Request) {
     try {
+      if(req['user'].role != "ADMIN" && req['user'].id == id){
+        throw new BadRequestException("You cannot delete othes account! Only ADMIN or SUPER-ADMIN can!")
+      }
       let deleted = await this.prisma.user.delete({where: {id}})
+      return deleted
     } catch (error) {
-      return {message: error.message}
+      throw new BadRequestException({message: error.message})
     }
   }
 }
